Fix efifatimagesize type
The efifatimagesize attribute type value is set to "oem" but
the documentation says that it is intended to be used for
creating ISO images.
This causes a schema error when this attribute is set on a profile with
type "iso" and blocks changing the EFI boot image size which is a problem
if the image is bigger than 20M.

This patch changes the efifatimagesize attribute type to "iso" with the benefit
of being able to change the EFI boot image size.